requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-dotenv
python-multipart
sqlalchemy
pymysql